datapath: Recreate CT entry without removing stale equivalent
Previously, when a stale CT entry (of non CT_SERVICE type) was detected
(from a previous connection to a service with the same IP/port, but
a different rev_nat_index), the entry was removed, and only afterwards
recreated.

The removal included the removal of related SNAT entries. However, in
such case there were no SNAT entries created, thus the removal of CT +
SNAT before the recreation was redundant.

Signed-off-by: Martynas Pumputis <m@lambda.lt>